Improved Character ID Management in Session
- Added _characterId in WorldSession to track the current character
- Use of _characterId in logs and chat packets
- Reset of _characterId upon logout
- Changed the guid type to DeleteCharacter (uint64_t)
- Fixed ID retrieval during character creation
- Sequential sending of login packets without co_return
- Annotation [[maybe_unused]] to avoid warnings about unused variables

+++ src/common/Database/include/Database/Char/CharWrapper.h
@@ -36,7 +36,7 @@ namespace Fireland::Database::Char
         Fireland::Utils::Async::async<std::optional<characters>> GetCharacterByGuid(uint32_t guid) noexcept;
         Fireland::Utils::Async::async<std::vector<characters>> GetCharactersForAccount(uint32_t accountid) noexcept;
         Fireland::Utils::Async::async<std::optional<characters>> CreateCharacter(characters c) noexcept;
-        Fireland::Utils::Async::async<bool> DeleteCharacter(uint32_t guid, uint32_t accountid) noexcept;
+        Fireland::Utils::Async::async<bool> DeleteCharacter(uint64_t guid, uint32_t accountid) noexcept;
     private:
         CharWrapper(boost::asio::any_io_executor executor, connection_pool_wrapper_options options) noexcept;
